Moved AlarmManagerBroadCastReceiver from models folder to services folder Fixed "Tasks are not sorted by priority" Changed the size of "Delete Button" and "Priority Image" bigger
