+++ app/src/main/java/uofm/software_engineering/group7/to_do_bot/services/AlarmManagerBroadcastReceiver.java
@@ -1,4 +1,4 @@
-package uofm.software_engineering.group7.to_do_bot.models;
+package uofm.software_engineering.group7.to_do_bot.services;
 
 import android.app.AlarmManager;
 import android.app.PendingIntent;
